First module among several that will allow for flexible processing pipelines.
In overview, the WebSocket dispatcher will receive commands and
will have to set up multi-stage processing pipelines in some cases.
For example, to receive a file over an external network connection
and report on its SHA1 sum, we might have a pipeline like:

 external socket reader -> decryptor -> hasher

Based on the current code, we can assume that the final stage in
the process will emit an event or send a network message on
completion.

There are two ways to implement this kind of processing pipeline
within Mojolicious:

1. write each processing element (decryptor, hasher) as a function
   or class that does I/O via some sort of IPC, preferably wrapped
   up in a Mojo::IOLoop::Stream object. (so it uses $in->on(read
   => sub { ...}) to read, and $out->write(...) to pass output.

2. turn each processing element into an EventEmitter, and hook
   them up together by having the consumer side subscribe to
   the "read" events emitted by the producer.

I'm plumping for the second approach here, even though it breaks
down if you need sub-processes (events can't be passed between
processes, unless you hook into the $sub->on(progress => ...)
handler.

If you have long processing pipelines, it's probably better to use
the first approach since the main event loop won't stall for as
long if all the steps are very CPU bound (meaning that other
asynchronous stuff should get a fair share of time) but I
don't think that will be a problem with my app. I will have
relatively short pipelines and even if I'm accepting multiple
requests at once (eg, requests to split new files in parallel),
the event loop will be able to service them without too much
latency.

Note that emitting an event and processing a subscribed callback
does all happen in the same event loop "cycle", so if it turns
out that a long chain of CPU-intensive callbacks is causing
problems with overall responsiveness, it's possible to return
control back to the event loop by changing "emit" events within
one of the processing elements to occur in the next event loop
tick (or "cycle", as I called it above).

If things get really bad, you can also run stuff in a sub-process,
though you'll probably need to use the IPC method described
above to interact with it.