Adding some documentation, but also breaking the code.
I'm going to refactor the code a bit and also add a new
parameter for optional parameters that will be stashed
in the object when it's created.

Later on, I might actually use this as an alternative to
creating new filter/sink classes, since I can use it to
avoid a lot of boilerplate code. The new classes would
simply pass all constructor parameter on to this class,
and re-bless the object with the correct class name.