Github doesn't have very good integration for wiki...
I created the wiki repo as a sub-tree of the main (code) repo with:

    git subtree add --prefix wiki https://github.com/declanmalone/App-IDA-Daemon.wiki.git master

However, the subtree needs to be pushed separately. In my case,
I had to add a new remote:

    git remote add wiki https://github.com/declanmalone/App-IDA-Daemon.wiki.git

But this can't be pushed to with a regular git push, so you have to use something
like what's in this script:

    git subtree push --prefix wiki wiki master

The first wiki is the directory, so you have to be in top-level, and the
second one is the remote that I added above.

Such crap!