Drop ioloop parameter from sink.
I don't need the functionality I could get from passing in a
Mojo::IOLoop object, and actually it would complicate the code
because I'd have to have different a start/stop implementation
depending on whether we're using the main event loop or a private
one. It's easiest if I just use the main event loop and keep
track of my {running} state variable and next_tick() to restart.