Rethinking how to refactor the Source, Filter and Sink classes:
* Base Link class that provides constructor and brings in
  Mojo::EventEmitter as its parent.

* A Role Filter that encapsulates most of the code to do filtering

* A class Link::NullFilter that inherits from Link and does Filter,
  providing the required filter_process() method.

The test script is just an experiment in how to compose roles into
a class:

* Using NullFilter, which uses inheritance and Roles

* Composing a new class manually with Mojo::Base->with_roles

From here, I will add simple Source and Sink roles, as well as
Split and Combine roles, together with concrete classes like
NullFilter that provide the concrete methods required by the
roles.

After I'm happy with the general organisation of the code and
that it works as expected, I will refactor to use sub-roles
to eliminate duplication across the various existing roles.