Repeated reads seem to be doing OK.
ioloop is not blocking after any of the reads.

Maybe I'm just misunderstanding how promises interact with the
event loop. The event loop probably just returns even if there are
promises outstanding. I guess that it's only in the case where
we have an explicit wait on a promise that the loop will block.

It would be nice to be able to trace promises that are still active
after returning from each call...